Refactor authentication into dedicated module
